Move components classes to specified package directory
